refactor: common의 Image를 테이블에 의존하지 않도록 수정
